+++ src/plugin.cpp
@@ -72,6 +72,7 @@ void init(rack::Plugin* p) {
 	p->addModel(modelAudioInterface64);
 	p->addModel(modelMb);
 	p->addModel(modelMe);
+	p->addModel(modelReel);
 
 	StoermelderPackOne::pluginSettings.readFromJson();
 	StoermelderPackOne::Ahab::Midi::init();