+++ src/plugin.hpp
@@ -77,6 +77,7 @@ extern Model* modelAhab;
 extern Model* modelAudioInterface64;
 extern Model* modelMb;
 extern Model* modelMe;
+extern Model* modelReel;
 #else
 extern Model* modelBolt;
 extern Model* modelFourRounds;